requirements.txt
psycopg2-binary
ortools
pandas
numpy
fastapi
uvicorn
sqlalchemy
python-dotenv
colorlog

